langinfo: pure-Python system for managing lang info (i.e. content types like 'Perl', 'XML', etc.) with a focus on text languages.
For starters this will be used for coming find/replace system additions, but
eventually will be used by codeintel and the ko*Language language services to
share basic static language info (e.g. the set of keywords).

Currently this doesn't not cover all languages defined in src/languages/..., but
it does define some that are not there as well.
(TODO: add all langs in src/languages to langinfo.)

r=shanec,toddw


git-svn-id: http://svn.openkomodo.com/repos/openkomodo/trunk/src/python-sitelib@536 b87d16e7-d29f-4ee1-b5ec-817d645d335f